# openapi/generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DataStack API",
    "version": "2.0.0",
    "description": "REST API for DataStack: authentication, workspaces, compute clusters, jobs, notebooks, pipelines, SQL warehouses, and webhooks."
  },
...
  "paths": {
...
    "/v1/pipelines": {
      "get": {
        "summary": "List pipelines",
        "tags": [
          "Pipelines"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "IDLE",
                "RUNNING",
                "FAILED",
                "STOPPING",
                "DELETED"
              ]
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25
            }
          },
          {
            "name": "pageToken",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pipelines": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "workspaceId": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "IDLE",
                              "RUNNING",
                              "FAILED",
                              "STOPPING",
                              "DELETED"
                            ]
                          },
                          "target": {
                            "type": "string",
                            "enum": [
                              "DEVELOPMENT",
                              "PRODUCTION"
                            ]
                          },
                          "catalog": {
                            "type": "string"
                          },
                          "schema": {
                            "type": "string"
                          },
                          "continuous": {
                            "type": "boolean"
                          },
                          "photon": {
                            "type": "boolean"
                          },
                          "edition": {
                            "type": "string",
                            "enum": [
                              "CORE",
                              "PRO",
                              "ADVANCED"
                            ]
                          },
                          "clusters": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "label": {
                                  "type": "string"
                                },
                                "nodeType": {
                                  "type": "string"
                                },
                                "workerCount": {
                                  "type": "integer"
                                },
                                "enableAutoscaling": {
                                  "type": "boolean"
                                }
                              }
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdBy": {
                            "type": "string"
                          },
                          "lastRunAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "workspaceId",
                          "state",
                          "target",
                          "createdAt"
                        ]
                      }
                    },
                    "totalCount": {
                      "type": "number"
                    },
                    "nextPageToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "pipelines",
                    "totalCount"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create pipeline",
        "tags": [
          "Pipelines"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "workspaceId": {
                    "type": "string"
                  },
                  "target": {
                    "type": "string",
                    "enum": [
                      "DEVELOPMENT",
                      "PRODUCTION"
                    ]
                  },
                  "catalog": {
                    "type": "string"
                  },
                  "schema": {
                    "type": "string"
                  },
                  "continuous": {
                    "type": "boolean"
                  },
                  "photon": {
                    "type": "boolean"
                  },
                  "edition": {
                    "type": "string",
                    "enum": [
                      "CORE",
                      "PRO",
                      "ADVANCED"
                    ]
                  },
                  "clusters": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "label": {
                          "type": "string"
                        },
                        "nodeType": {
                          "type": "string"
                        },
                        "workerCount": {
                          "type": "integer"
                        },
                        "enableAutoscaling": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "name",
                  "workspaceId",
                  "target"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "IDLE",
                        "RUNNING",
                        "FAILED",
                        "STOPPING",
                        "DELETED"
                      ]
                    },
                    "target": {
                      "type": "string",
                      "enum": [
                        "DEVELOPMENT",
                        "PRODUCTION"
                      ]
                    },
                    "catalog": {
                      "type": "string"
                    },
                    "schema": {
                      "type": "string"
                    },
                    "continuous": {
                      "type": "boolean"
                    },
                    "photon": {
                      "type": "boolean"
                    },
                    "edition": {
                      "type": "string",
                      "enum": [
                        "CORE",
                        "PRO",
                        "ADVANCED"
                      ]
                    },
                    "clusters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string"
                          },
                          "nodeType": {
                            "type": "string"
                          },
                          "workerCount": {
                            "type": "integer"
                          },
                          "enableAutoscaling": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "lastRunAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "workspaceId",
                    "state",
                    "target",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/pipelines/{pipelineId}": {
      "get": {
        "summary": "Get pipeline by ID",
        "tags": [
          "Pipelines"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "IDLE",
                        "RUNNING",
                        "FAILED",
                        "STOPPING",
                        "DELETED"
                      ]
                    },
                    "target": {
                      "type": "string",
                      "enum": [
                        "DEVELOPMENT",
                        "PRODUCTION"
                      ]
                    },
                    "catalog": {
                      "type": "string"
                    },
                    "schema": {
                      "type": "string"
                    },
                    "continuous": {
                      "type": "boolean"
                    },
                    "photon": {
                      "type": "boolean"
                    },
                    "edition": {
                      "type": "string",
                      "enum": [
                        "CORE",
                        "PRO",
                        "ADVANCED"
                      ]
                    },
                    "clusters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string"
                          },
                          "nodeType": {
                            "type": "string"
                          },
                          "workerCount": {
                            "type": "integer"
                          },
                          "enableAutoscaling": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "lastRunAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "workspaceId",
                    "state",
                    "target",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete pipeline",
        "tags": [
          "Pipelines"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/pipelines/{pipelineId}/start": {
      "post": {
        "summary": "Start pipeline",
        "tags": [
          "Pipelines"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullRefresh": {
                    "type": "boolean"
                  },
                  "refreshSelection": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "IDLE",
                        "RUNNING",
                        "FAILED",
                        "STOPPING",
                        "DELETED"
                      ]
                    },
                    "target": {
                      "type": "string",
                      "enum": [
                        "DEVELOPMENT",
                        "PRODUCTION"
                      ]
                    },
                    "catalog": {
                      "type": "string"
                    },
                    "schema": {
                      "type": "string"
                    },
                    "continuous": {
                      "type": "boolean"
                    },
                    "photon": {
                      "type": "boolean"
                    },
                    "edition": {
                      "type": "string",
                      "enum": [
                        "CORE",
                        "PRO",
                        "ADVANCED"
                      ]
                    },
                    "clusters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string"
                          },
                          "nodeType": {
                            "type": "string"
                          },
                          "workerCount": {
                            "type": "integer"
                          },
                          "enableAutoscaling": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "lastRunAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "workspaceId",
                    "state",
                    "target",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/pipelines/{pipelineId}/events": {
      "get": {
        "summary": "List pipeline events",
        "tags": [
          "Pipelines"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "timestamp_asc",
                "timestamp_desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "pipelineId": {
                            "type": "string"
                          },
                          "eventType": {
                            "type": "string",
                            "enum": [
                              "STARTED",
                              "COMPLETED",
                              "FAILED",
                              "RETRYING"
                            ]
                          },
                          "message": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "details": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "id",
                          "pipelineId",
                          "eventType",
                          "timestamp"
                        ]
                      }
                    },
                    "totalCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "events",
                    "totalCount"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}